feat: buildType によって配布する実装を変更
