+++ app/build.gradle.kts
@@ -25,6 +25,11 @@ android {
       isMinifyEnabled = false
       proguardFiles(getDefaultProguardFile("proguard-android-optimize.txt"), "proguard-rules.pro")
     }
+
+    create("mock") {
+      initWith(getByName("debug"))
+      matchingFallbacks += listOf("debug")
+    }
   }
 
   compileOptions {
