+++ core/data/build.gradle.kts
@@ -26,6 +26,11 @@ android {
       isMinifyEnabled = false
       proguardFiles(getDefaultProguardFile("proguard-android-optimize.txt"), "proguard-rules.pro")
     }
+
+    create("mock") {
+      initWith(getByName("debug"))
+      matchingFallbacks += listOf("debug")
+    }
   }
 
   compileOptions {
