+++ core/data/src/debug/java/app/doggy/mybrary/core/data/BookRepositoryModule.kt
@@ -1,0 +1,18 @@
+package app.doggy.mybrary.core.data
+
+import app.doggy.mybrary.core.data.repository.BookRepositoryImpl
+import app.doggy.mybrary.core.domain.repository.BookRepository
+import dagger.Binds
+import dagger.Module
+import dagger.hilt.InstallIn
+import dagger.hilt.components.SingletonComponent
+
+@Module
+@InstallIn(SingletonComponent::class)
+internal interface BookRepositoryModule {
+
+  @Binds
+  fun bindsBookRepository(
+    bookRepository: BookRepositoryImpl,
+  ): BookRepository
+}
