+++ core/data/src/debug/java/app/doggy/mybrary/core/data/DraftMemoRepositoryModule.kt
@@ -1,0 +1,18 @@
+package app.doggy.mybrary.core.data
+
+import app.doggy.mybrary.core.data.repository.DraftMemoRepositoryImpl
+import app.doggy.mybrary.core.domain.repository.DraftMemoRepository
+import dagger.Binds
+import dagger.Module
+import dagger.hilt.InstallIn
+import dagger.hilt.components.SingletonComponent
+
+@Module
+@InstallIn(SingletonComponent::class)
+internal interface DraftMemoRepositoryModule {
+
+  @Binds
+  fun bindsDraftMemoRepository(
+    draftMemoRepository: DraftMemoRepositoryImpl,
+  ): DraftMemoRepository
+}
