+++ core/data/src/debug/java/app/doggy/mybrary/core/data/MyBookRepositoryModule.kt
@@ -1,0 +1,18 @@
+package app.doggy.mybrary.core.data
+
+import app.doggy.mybrary.core.data.repository.MyBookRepositoryImpl
+import app.doggy.mybrary.core.domain.repository.MyBookRepository
+import dagger.Binds
+import dagger.Module
+import dagger.hilt.InstallIn
+import dagger.hilt.components.SingletonComponent
+
+@Module
+@InstallIn(SingletonComponent::class)
+internal interface MyBookRepositoryModule {
+
+  @Binds
+  fun bindsMyBookRepository(
+    myBookRepository: MyBookRepositoryImpl,
+  ): MyBookRepository
+}
