+++ core/data/src/main/java/app/doggy/mybrary/core/data/di/BookRepositoryModule.kt
@@ -7,12 +7,12 @@ import dagger.Module
 import dagger.hilt.InstallIn
 import dagger.hilt.components.SingletonComponent
 
-@Module
-@InstallIn(SingletonComponent::class)
-internal interface BookRepositoryModule {
-
-  @Binds
-  fun bindsBookRepository(
-    bookRepository: BookRepositoryImpl,
-  ): BookRepository
-}
+// @Module
+// @InstallIn(SingletonComponent::class)
+// internal interface BookRepositoryModule {
+//
+//   @Binds
+//   fun bindsBookRepository(
+//     bookRepository: BookRepositoryImpl,
+//   ): BookRepository
+// }
