+++ core/data/src/main/java/app/doggy/mybrary/core/data/di/MemoRepositoryModule.kt
@@ -7,12 +7,12 @@ import dagger.Module
 import dagger.hilt.InstallIn
 import dagger.hilt.components.SingletonComponent
 
-@Module
-@InstallIn(SingletonComponent::class)
-internal interface MemoRepositoryModule {
-
-  @Binds
-  fun bindsMemoRepository(
-    memoRepository: MemoRepositoryImpl,
-  ): MemoRepository
-}
+// @Module
+// @InstallIn(SingletonComponent::class)
+// internal interface MemoRepositoryModule {
+//
+//   @Binds
+//   fun bindsMemoRepository(
+//     memoRepository: MemoRepositoryImpl,
+//   ): MemoRepository
+// }
