+++ core/data/src/main/java/app/doggy/mybrary/core/data/di/MyBookRepositoryModule.kt
@@ -7,12 +7,12 @@ import dagger.Module
 import dagger.hilt.InstallIn
 import dagger.hilt.components.SingletonComponent
 
-@Module
-@InstallIn(SingletonComponent::class)
-internal interface MyBookRepositoryModule {
-
-  @Binds
-  fun bindsMyBookRepository(
-    myBookRepository: MyBookRepositoryImpl,
-  ): MyBookRepository
-}
+// @Module
+// @InstallIn(SingletonComponent::class)
+// internal interface MyBookRepositoryModule {
+//
+//   @Binds
+//   fun bindsMyBookRepository(
+//     myBookRepository: MyBookRepositoryImpl,
+//   ): MyBookRepository
+// }
