--- core/data/src/main/java/app/doggy/mybrary/core/data/mock/di/DebugDraftMemoRepositoryModule.kt
@@ -1,18 +1,0 @@
-package app.doggy.mybrary.core.data.di
-
-import app.doggy.mybrary.core.data.repository.DraftMemoRepositoryImpl
-import app.doggy.mybrary.core.domain.repository.DraftMemoRepository
-import dagger.Binds
-import dagger.Module
-import dagger.hilt.InstallIn
-import dagger.hilt.components.SingletonComponent
-
-// @Module
-// @InstallIn(SingletonComponent::class)
-// internal interface DebugDraftMemoRepositoryModule {
-//
-//   @Binds
-//   fun bindsDraftMemoRepository(
-//     draftMemoRepository: DraftMemoRepositoryImpl,
-//   ): DraftMemoRepository
-// }
