--- core/data/src/main/java/app/doggy/mybrary/core/data/mock/di/DebugMyBookRepositoryModule.kt
@@ -1,18 +1,0 @@
-package app.doggy.mybrary.core.data.di
-
-import app.doggy.mybrary.core.data.repository.MyBookRepositoryImpl
-import app.doggy.mybrary.core.domain.repository.MyBookRepository
-import dagger.Binds
-import dagger.Module
-import dagger.hilt.InstallIn
-import dagger.hilt.components.SingletonComponent
-
-// @Module
-// @InstallIn(SingletonComponent::class)
-// internal interface DebugMyBookRepositoryModule {
-//
-//   @Binds
-//   fun bindsMyBookRepository(
-//     myBookRepository: MyBookRepositoryImpl,
-//   ): MyBookRepository
-// }
