+++ core/data/src/mock/java/app/doggy/mybrary/core/data/model/Memo.kt
